Auto stash before merge of "master" and "origin/master"

+++ Arkanoid/src/app/BonusDuplicate.java
@@ -35,7 +35,7 @@ public class BonusDuplicate extends Bonus {
     }
 
     @Override
-    public int aceleracion() {
+    public double aceleracion() {
         return 0;
     }
